+++ lihil/websocket.py
@@ -34,21 +34,22 @@ class WebSocketEndpoint:  # TODO:  endpoint base
     def props(self):
         return self._props
 
-    async def chainup_plugins(
+    def chainup_plugins(
         self, func: Callable[..., Awaitable[None]], sig: EndpointSignature[None]
     ) -> Callable[..., Awaitable[None]]:
+        seen: set[int] = set()
         for decor in self._props.plugins:
-            if iscoroutinefunction(decor):
-                wrapped = await decor(self._route.graph, func, sig)
-            else:
-                wrapped = decor(self._route.graph, func, sig)
+            if (decor_id := id(decor)) in seen:
+                continue
+            wrapped = decor(self._route.graph, func, sig)
             func = cast(Callable[..., Awaitable[None]], wrapped)
+            seen.add(decor_id)
         return func
 
-    async def setup(self, sig: EndpointSignature[None]) -> None:
+    def setup(self, sig: EndpointSignature[None]) -> None:
         self._graph = self._route.graph
         self._sig = sig
-        self._func = await self.chainup_plugins(self._func, sig)
+        self._func = self.chainup_plugins(self._func, sig)
         self._injector = Injector(self._sig)
         self._scoped: bool = self._sig.scoped
 
@@ -90,26 +91,26 @@ class WebSocketRoute(RouteBase):
 
     async def __call__(self, scope: IScope, receive: IReceive, send: ISend) -> None:
         if not self.call_stack:
-            raise RuntimeError(f"{self.__class__.__name__}({self.path}) not setup")
+            raise RuntimeError(f"{self.__class__.__name__}({self._path}) not setup")
         await self.call_stack(scope, receive, send)
 
     def __repr__(self):
-        return f"{self.__class__.__name__}({self.path!r}, {self.endpoint})"
+        return f"{self.__class__.__name__}({self._path!r}, {self.endpoint})"
 
-    async def setup(
+    def _setup(
         self, graph: Graph | None = None, workers: ThreadPoolExecutor | None = None
     ):
         if self.endpoint is None:
             raise RuntimeError(f"Empty websocket route")
 
-        await super().setup(graph=graph, workers=workers)
-        self.endpoint_parser = EndpointParser(self.graph, self.path)
+        super()._setup(graph=graph, workers=workers)
+        self.endpoint_parser = EndpointParser(self._graph, self._path)
         sig = self.endpoint_parser.parse(self.endpoint.unwrapped_func)
         if sig.body_param is not None:
             raise NotSupportedError(
                 f"Websocket does not support body param, got {sig.body_param}"
             )
-        await self.endpoint.setup(sig)
+        self.endpoint.setup(sig)
         self.call_stack = self.chainup_middlewares(self.endpoint)
 
     def include_subroutes(self, *subs: Self, parent_prefix: str | None = None) -> None:
@@ -120,24 +121,24 @@ class WebSocketRoute(RouteBase):
         NOTE: This method is NOT idempotent
         """
         for sub in subs:
-            self.graph.merge(sub.graph)
+            self._graph.merge(sub._graph)
             if parent_prefix:
-                sub_path = sub.path.removeprefix(parent_prefix)
+                sub_path = sub._path.removeprefix(parent_prefix)
             else:
-                sub_path = sub.path
-            merged_path = merge_path(self.path, sub_path)
-            sub_subs = sub.subroutes
+                sub_path = sub._path
+            merged_path = merge_path(self._path, sub_path)
+            sub_subs = sub._subroutes
             new_sub = self.__class__(
                 path=merged_path,
-                graph=self.graph,
+                graph=self._graph,
                 middlewares=sub.middle_factories,
             )
             if sub.endpoint is not None:
                 new_sub.ws_handler(sub.endpoint.unwrapped_func, **sub.endpoint.props)
 
             for sub_sub in sub_subs:
-                new_sub.include_subroutes(sub_sub, parent_prefix=sub.path)
-            self.subroutes.append(new_sub)
+                new_sub.include_subroutes(sub_sub, parent_prefix=sub._path)
+            self._subroutes.append(new_sub)
 
     def ws_handler(self, func: Any = None, **iprops: Unpack[IEndpointProps]) -> Any:
         props = EndpointProps.from_unpack(**iprops)