+++ tests/test_endpoint/test_resp_mark.py
@@ -26,7 +26,6 @@ async def test_endpoint_deps():
     route = Route()
     route.get(get_order)
     ep = route.get_endpoint("GET")
-    await route.setup()
     rt = ep.sig.return_params[200]
     assert rt.type_ == Union[Order, str]