+++ tests/test_plugins/test_auth.py
@@ -28,7 +28,6 @@ async def test_login():
     token.post(create_token)
 
     form_ep = token.get_endpoint("POST")
-    await token.setup()
 
     lc = LocalClient()
     res = await lc.submit_form(